Application/Application/7 Docker Compose files/requirements.txt
Flask==2.3.2
prometheus_client==0.17.0
requests==2.31.0

opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0
